requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.1
